refactor: changes as per review
